pyproject.toml
[project]
name = "kinitro-agent"
version = "0.0.1"
description = "Kinitro Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "metaworld>=3.0.0",
    "torch>=2.8.0",
    "gymnasium>=0.29.0",
    "mujoco>=3.0.0",
    "numpy>=1.24.0",
    "pycapnp>=2.1.0",
    "tensorboard>=2.15.0",
    "matplotlib>=3.7.0"
]

[dependency-groups]
dev = ["debugpy>=1.8.9", "py-spy>=0.4.0", "pytest>=8.3.4", "ruff>=0.8.2"]
